Improve filaname handling and manifest handling. Change interface to Manifest.write to take a directory (or s3path) an a filename separately.
Add test for generating filenames automatically for manifest files

Add a validation method to Manifest class to optionally check file checksums on demand.

Add some tests for LiberaProductFilename

Add a usedforsecurity=False kwarg to md5 checksum.

Rework filename class interfaces to better support pathlib and cloudpathlib

Add a test for messing with the path attribute on an instance of AbstractValidFilename.
Move the regex check to before we set the path attribute on an instance.

Add function to filenaming module for getting the current formatted revision based on the current time.
Add function to filenaming module for formatting a semantic version into a filename friendly string.